refactored step3 to have custom translation components and only 2 addresses. Fixed ui
